back/yolo/requirements.txt
uvicorn
fastapi
pydantic
ultralytics
pillow
eva-decord
python-multipart